+++ src/helpers/collection_helpers.ts
@@ -10,9 +10,11 @@ import { isAccessTokenRevokedError } from '../api/api';
 import { isNumberOrNumberString } from "./math_helper";
 import { Logger } from './logger';
 import { sleep } from './time_helpers';
+import { UNIQUE_ID_PROPERTIES, type ApiCacheKey } from "../enums";
 import {
+    type ApiObj,
     type CountKey,
-    type MastodonObjWithID,
+    type ApiObjWithID,
     type MinMax,
     type MinMaxID,
     type OptionalNumber,
@@ -131,16 +133,14 @@ export async function batchMap<T, U>(
 
 /**
  * Checks if the elements of an array have unique IDs and logs a warning if not.
- * @param {MastodonObjWithID[]} array - Array of objects with IDs.
+ * @param {ApiObjWithID[]} array - Array of objects with IDs.
  * @param {Logger} logger - Logger to use for warnings.
  */
-export function checkUniqueIDs(array: MastodonObjWithID[], logger: Logger): void {
-    const objsById = groupBy<MastodonObjWithID>(array, (e) => e.id);
-    const uniqueIds = Object.keys(objsById);
+export function checkUniqueRows<T extends ApiObj>(cacheKey: ApiCacheKey, array: T[], logger: Logger): void {
+    const uniqObjs = uniquifyApiObjs(cacheKey, array, logger);
 
-    if (uniqueIds.length != array.length) {
-        const objsWithDuplicates = Object.entries(objsById).filter(([_, objs]) => objs.length > 1);
-        logger.warn(`${array.length} objs only have ${uniqueIds.length} unique IDs! Dupes:`, objsWithDuplicates);
+    if (uniqObjs.length != array.length) {
+        logger.warn(`checkUniqueRows() Found ${array.length - uniqObjs.length} duplicate objects in "${cacheKey}"`);
     }
 };
 
@@ -230,10 +230,10 @@ export function filterWithLog<T>(
  *     const idx = Math.min(toots.length - 1, MAX_ID_IDX);
  *     return sortByCreatedAt(toots)[idx].id;
  * }
- * @param {MastodonObjWithID[]} array - Array of objects with IDs.
+ * @param {ApiObjWithID[]} array - Array of objects with IDs.
  * @returns {MinMaxID | null} The min and max IDs, or null if invalid.
  */
-export function findMinMaxId(array: MastodonObjWithID[]): MinMaxID | null {
+export function findMinMaxId(array: ApiObjWithID[]): MinMaxID | null {
     if (!array?.length) {
         console.warn(`[findMinMaxId()] called with 0 length array:`, array);
         return null;
@@ -290,7 +290,7 @@ export async function getPromiseResults<T>(promises: Promise<T>[]): Promise<Prom
  * @param {(item: T) => string} makeKey - Function to get group key.
  * @returns {Record<string, T[]>} The grouped object.
  */
-export function groupBy<T>(array: T[], makeKey: (item: T) => string): Record<string, T[]> {
+export function groupBy<T>(array: T[], makeKey: (item: T) => string | number): Record<string, T[]> {
     return array.reduce(
         (grouped, item) => {
             const group = makeKey(item);
@@ -345,7 +345,7 @@ export function decrementCount(counts: StringNumberDict, k?: CountKey | null, in
  * @param {T[]} array - Array of objects with id property.
  * @returns {Record<string, T>} The keyed dictionary.
  */
-export function keyById<T extends MastodonObjWithID>(array: T[]): Record<string, T> {
+export function keyById<T extends ApiObjWithID>(array: T[]): Record<string, T> {
     return keyByProperty<T>(array, obj => obj.id);
 };
 
@@ -709,6 +709,31 @@ export const uniquify = (array: (string | undefined)[]): string[] | undefined =>
     let newArray = array.filter((e) => e != undefined) as string[];
     newArray = [...new Set(newArray)];
     return newArray;
+};
+
+
+/**
+ * Uniquify an array of API objects by the appropriate property. This is a no-op for API objects
+ * that don't have a property that can be used to uniquely identify them.
+ * @template T
+ * @param {ApiCacheKey} cacheKey - The cache key to determine the unique property.
+ * @param {T[]} array - Array of API objects.
+ * @param {Logger} logger - Logger to use for warnings.
+ */
+export function uniquifyApiObjs<T extends ApiObj>(cacheKey: ApiCacheKey, array: T[], logger: Logger): T[] {
+    const uniqueProperty = UNIQUE_ID_PROPERTIES[cacheKey] as keyof ApiObj;
+    const thisLogger = logger.tempLogger(`uniquifyApiObjs`);
+
+    if (!uniqueProperty) {
+        thisLogger.trace(`No unique property for "${cacheKey}", skipping uniquify...`);
+        return array;
+    } else if (array.length && isNil(array[0][uniqueProperty])) {
+        thisLogger.error(`checkUniqueRows() called with array that has no "${uniqueProperty}" property!`, array);
+        return array;
+    }
+
+    logger.trace(`Uniquifying array of ${array.length} objects by "${uniqueProperty}" property`);
+    return uniquifyByProp(array, (obj) => obj[uniqueProperty], cacheKey);
 };